Fixed encoding problem in Lukas code. (we should not use diacritics in code. If you need to do it, please don't use cp1250 - use utf8 instead) Fixed signed with unsigned comparsion warning in Hardwire code.

+++ src/EntitySystem/EntityMgr/EntityPicker.cpp
@@ -19,7 +19,7 @@ public:
 		return true;
 	}
 
-	static const size_t MAX_QUERY_SHAPES = 128;
+	static const int32 MAX_QUERY_SHAPES = 128;
 	PhysicalShape* shapes[MAX_QUERY_SHAPES];
 	int32 count;
 };
